Move and package api.go and api_test.go in an 'api' folder
Moved and packaged the api code to an api package so that it can be
referenced by other go projects (such as the serverless golang
starter snake).
